Merge from Master

+++ BookStore/BookStore/Models/Cart.cs
@@ -17,25 +17,25 @@ namespace BookStore.Models
         public string cartUserID { get; set; }
 
         [Display(Name = "SubTotal:")]
-        public decimal? subTotal { get; set; }
+        public double? subTotal { get; set; }
 
         [Display(Name = "GST:")]
-        public decimal? gst { get; set; }
+        public double? gst { get; set; }
 
         [Display(Name = "Discount Percent:")]
-        public decimal? discountPercent { get; set; }
+        public double? discountPercent { get; set; }
 
         [Display(Name = "Discount Amount:")]
-        public decimal? discountAmount { get; set; }
+        public double? discountAmount { get; set; }
 
         [Display(Name = "Delivery Charge:")]
-        public decimal? deliveryCharge { get; set; }
+        public double? deliveryCharge { get; set; }
 
         [Display(Name = "Collection Day:")]
         public int? collectionDay { get; set; }
 
         [Display(Name = "Total Price:")]
-        public decimal? totalPrice { get; set; }
+        public double? totalPrice { get; set; }
 
         [Display(Name = "Date of Purchase:")]
         public DateTime? dateOfPurchase { get; set; }
